+++ sdk/storage/azure-storage-files-shares/src/share_directory_client.cpp
@@ -134,17 +134,21 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       protocolLayerOptions.FileAttributes
           = Details::FileAttributesToString(Models::FileAttributes::Directory);
     }
-    if (options.SmbProperties.CreationTime.HasValue())
+    if (options.SmbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime = options.SmbProperties.CreationTime.GetValue();
+      protocolLayerOptions.FileCreationTime
+          = options.SmbProperties.CreatedOn.GetValue().GetRfc3339String(
+              Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
       protocolLayerOptions.FileCreationTime = std::string(c_FileDefaultTimeValue);
     }
-    if (options.SmbProperties.LastWriteTime.HasValue())
+    if (options.SmbProperties.LastWrittenOn.HasValue())
     {
-      protocolLayerOptions.FileLastWriteTime = options.SmbProperties.LastWriteTime.GetValue();
+      protocolLayerOptions.FileLastWriteTime
+          = options.SmbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+              Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
@@ -188,21 +192,24 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   {
     auto protocolLayerOptions = Details::ShareRestClient::Directory::SetPropertiesOptions();
     protocolLayerOptions.FileAttributes = Details::FileAttributesToString(smbProperties.Attributes);
-    if (smbProperties.CreationTime.HasValue())
+    if (smbProperties.CreatedOn.HasValue())
     {
-      protocolLayerOptions.FileCreationTime = smbProperties.CreationTime.GetValue();
+      protocolLayerOptions.FileCreationTime = smbProperties.CreatedOn.GetValue().GetRfc3339String(
+          Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
-      protocolLayerOptions.FileCreationTime = std::string(c_FilePreserveSmbProperties);
+      protocolLayerOptions.FileCreationTime = std::string(c_FileDefaultTimeValue);
     }
-    if (smbProperties.LastWriteTime.HasValue())
+    if (smbProperties.LastWrittenOn.HasValue())
     {
-      protocolLayerOptions.FileLastWriteTime = smbProperties.LastWriteTime.GetValue();
+      protocolLayerOptions.FileLastWriteTime
+          = smbProperties.LastWrittenOn.GetValue().GetRfc3339String(
+              Core::DateTime::TimeFractionFormat::AllDigits);
     }
     else
     {
-      protocolLayerOptions.FileLastWriteTime = std::string(c_FilePreserveSmbProperties);
+      protocolLayerOptions.FileLastWriteTime = std::string(c_FileDefaultTimeValue);
     }
     if (options.FilePermission.HasValue())
     {